requirements.txt
discord.py
python-dotenv
firebase-admin
